LPS-136506 Moves code to RenderCollectionLayoutStructureItemDisplayContext, no logic changes
