+++ modules/apps/layout/layout-taglib/src/main/java/com/liferay/layout/taglib/internal/display/context/RenderLayoutStructureDisplayContext.java
@@ -15,7 +15,6 @@
 package com.liferay.layout.taglib.internal.display.context;
 
 import com.liferay.asset.info.display.contributor.util.ContentAccessor;
-import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.fragment.entry.processor.helper.FragmentEntryProcessorHelper;
 import com.liferay.fragment.model.FragmentEntryLink;
 import com.liferay.fragment.renderer.DefaultFragmentRendererContext;
@@ -31,24 +30,10 @@ import com.liferay.info.item.InfoItemReference;
 import com.liferay.info.item.InfoItemServiceTracker;
 import com.liferay.info.item.provider.InfoItemFieldValuesProvider;
 import com.liferay.info.item.provider.InfoItemObjectProvider;
-import com.liferay.info.list.renderer.DefaultInfoListRendererContext;
-import com.liferay.info.list.renderer.InfoListRenderer;
-import com.liferay.info.list.renderer.InfoListRendererContext;
-import com.liferay.info.list.renderer.InfoListRendererTracker;
-import com.liferay.info.pagination.Pagination;
 import com.liferay.info.type.WebImage;
-import com.liferay.layout.display.page.LayoutDisplayPageProvider;
-import com.liferay.layout.display.page.LayoutDisplayPageProviderTracker;
-import com.liferay.layout.list.retriever.DefaultLayoutListRetrieverContext;
-import com.liferay.layout.list.retriever.LayoutListRetriever;
-import com.liferay.layout.list.retriever.LayoutListRetrieverTracker;
-import com.liferay.layout.list.retriever.ListObjectReference;
-import com.liferay.layout.list.retriever.ListObjectReferenceFactory;
-import com.liferay.layout.list.retriever.ListObjectReferenceFactoryTracker;
 import com.liferay.layout.responsive.ResponsiveLayoutStructureUtil;
 import com.liferay.layout.taglib.internal.servlet.ServletContextUtil;
 import com.liferay.layout.util.constants.LayoutDataItemTypeConstants;
-import com.liferay.layout.util.structure.CollectionStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.ContainerStyledLayoutStructureItem;
 import com.liferay.layout.util.structure.LayoutStructure;
 import com.liferay.layout.util.structure.LayoutStructureItem;
@@ -78,46 +63,30 @@ import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
-import com.liferay.portal.util.PropsValues;
-import com.liferay.segments.SegmentsEntryRetriever;
 import com.liferay.segments.constants.SegmentsExperienceConstants;
 import com.liferay.segments.constants.SegmentsWebKeys;
-import com.liferay.segments.context.RequestContextMapper;
 import com.liferay.style.book.model.StyleBookEntry;
 import com.liferay.style.book.util.DefaultStyleBookEntryUtil;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * @author Rubén Pulido
  */
 public class RenderLayoutStructureDisplayContext {
 
-	public static final String PAGE_NUMBER_PARAM_PREFIX = "page_number_";
-
-	public static final String PAGINATION_TYPE_NUMERIC = "numeric";
-
-	public static final String PAGINATION_TYPE_SIMPLE = "simple";
-
 	public RenderLayoutStructureDisplayContext(
 		Map<String, Object> fieldValues, HttpServletRequest httpServletRequest,
-		HttpServletResponse httpServletResponse,
 		LayoutStructure layoutStructure, String mainItemId, String mode,
 		boolean showPreview) {
 
 		_fieldValues = fieldValues;
 		_httpServletRequest = httpServletRequest;
-		_httpServletResponse = httpServletResponse;
 		_layoutStructure = layoutStructure;
 		_mainItemId = mainItemId;
 		_mode = mode;
@@ -125,130 +94,6 @@ public class RenderLayoutStructureDisplayContext {
 
 		_themeDisplay = (ThemeDisplay)_httpServletRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
-	}
-
-	public List<Object> getCollection(
-		CollectionStyledLayoutStructureItem
-			collectionStyledLayoutStructureItem) {
-
-		JSONObject collectionJSONObject =
-			collectionStyledLayoutStructureItem.getCollectionJSONObject();
-
-		LayoutListRetriever<?, ListObjectReference> layoutListRetriever =
-			_getLayoutListRetriever(collectionJSONObject);
-		ListObjectReference listObjectReference = _getListObjectReference(
-			collectionJSONObject);
-
-		if ((layoutListRetriever == null) || (listObjectReference == null)) {
-			return Collections.emptyList();
-		}
-
-		DefaultLayoutListRetrieverContext defaultLayoutListRetrieverContext =
-			_getDefaultLayoutListRetrieverContext(collectionJSONObject);
-
-		int end = collectionStyledLayoutStructureItem.getNumberOfItems();
-		int start = 0;
-
-		String paginationType =
-			collectionStyledLayoutStructureItem.getPaginationType();
-
-		if (Objects.equals(paginationType, PAGINATION_TYPE_NUMERIC) ||
-			Objects.equals(paginationType, PAGINATION_TYPE_SIMPLE)) {
-
-			HttpServletRequest originalHttpServletRequest =
-				PortalUtil.getOriginalServletRequest(_httpServletRequest);
-
-			int currentPage = ParamUtil.getInteger(
-				originalHttpServletRequest,
-				PAGE_NUMBER_PARAM_PREFIX +
-					collectionStyledLayoutStructureItem.getItemId());
-
-			if (currentPage < 1) {
-				currentPage = 1;
-			}
-
-			int numberOfItems =
-				collectionStyledLayoutStructureItem.getNumberOfItems();
-
-			int numberOfItemsPerPage =
-				collectionStyledLayoutStructureItem.getNumberOfItemsPerPage();
-
-			if (numberOfItemsPerPage >
-					PropsValues.SEARCH_CONTAINER_PAGE_MAX_DELTA) {
-
-				numberOfItemsPerPage =
-					PropsValues.SEARCH_CONTAINER_PAGE_MAX_DELTA;
-			}
-
-			int listCount = layoutListRetriever.getListCount(
-				listObjectReference, defaultLayoutListRetrieverContext);
-
-			end = Math.min(
-				Math.min(currentPage * numberOfItemsPerPage, numberOfItems),
-				listCount);
-
-			start = (currentPage - 1) * numberOfItemsPerPage;
-		}
-
-		defaultLayoutListRetrieverContext.setPagination(
-			Pagination.of(end, start));
-
-		return layoutListRetriever.getList(
-			listObjectReference, defaultLayoutListRetrieverContext);
-	}
-
-	public int getCollectionCount(
-		CollectionStyledLayoutStructureItem
-			collectionStyledLayoutStructureItem) {
-
-		JSONObject collectionJSONObject =
-			collectionStyledLayoutStructureItem.getCollectionJSONObject();
-
-		LayoutListRetriever<?, ListObjectReference> layoutListRetriever =
-			_getLayoutListRetriever(collectionJSONObject);
-		ListObjectReference listObjectReference = _getListObjectReference(
-			collectionJSONObject);
-
-		if ((layoutListRetriever == null) || (listObjectReference == null)) {
-			return 0;
-		}
-
-		return layoutListRetriever.getListCount(
-			listObjectReference,
-			_getDefaultLayoutListRetrieverContext(collectionJSONObject));
-	}
-
-	public LayoutDisplayPageProvider<?> getCollectionLayoutDisplayPageProvider(
-		CollectionStyledLayoutStructureItem
-			collectionStyledLayoutStructureItem) {
-
-		JSONObject collectionJSONObject =
-			collectionStyledLayoutStructureItem.getCollectionJSONObject();
-
-		if ((collectionJSONObject == null) ||
-			(collectionJSONObject.length() <= 0)) {
-
-			return null;
-		}
-
-		ListObjectReference listObjectReference = _getListObjectReference(
-			collectionJSONObject);
-
-		if (listObjectReference == null) {
-			return null;
-		}
-
-		LayoutDisplayPageProviderTracker layoutDisplayPageProviderTracker =
-			ServletContextUtil.getLayoutDisplayPageProviderTracker();
-
-		String className = listObjectReference.getItemType();
-
-		if (Objects.equals(className, DLFileEntry.class.getName())) {
-			className = FileEntry.class.getName();
-		}
-
-		return layoutDisplayPageProviderTracker.
-			getLayoutDisplayPageProviderByClassName(className);
 	}
 
 	public String getContainerLinkHref(
@@ -619,36 +464,6 @@ public class RenderLayoutStructureDisplayContext {
 		return defaultFragmentRendererContext;
 	}
 
-	public InfoListRenderer<?> getInfoListRenderer(
-		CollectionStyledLayoutStructureItem
-			collectionStyledLayoutStructureItem) {
-
-		if (Validator.isNull(
-				collectionStyledLayoutStructureItem.getListStyle())) {
-
-			return null;
-		}
-
-		InfoListRendererTracker infoListRendererTracker =
-			ServletContextUtil.getInfoListRendererTracker();
-
-		return infoListRendererTracker.getInfoListRenderer(
-			collectionStyledLayoutStructureItem.getListStyle());
-	}
-
-	public InfoListRendererContext getInfoListRendererContext(
-		String listItemStyle, String templateKey) {
-
-		DefaultInfoListRendererContext defaultInfoListRendererContext =
-			new DefaultInfoListRendererContext(
-				_httpServletRequest, _httpServletResponse);
-
-		defaultInfoListRendererContext.setListItemRendererKey(listItemStyle);
-		defaultInfoListRendererContext.setTemplateKey(templateKey);
-
-		return defaultInfoListRendererContext;
-	}
-
 	public LayoutStructure getLayoutStructure() {
 		return _layoutStructure;
 	}
@@ -1014,63 +829,6 @@ public class RenderLayoutStructureDisplayContext {
 		return StringPool.BLANK;
 	}
 
-	private Map<String, String[]> _getConfiguration(
-		JSONObject collectionJSONObject) {
-
-		JSONObject configurationJSONObject = collectionJSONObject.getJSONObject(
-			"config");
-
-		if (configurationJSONObject == null) {
-			return null;
-		}
-
-		Map<String, String[]> configuration = new HashMap<>();
-
-		for (String key : configurationJSONObject.keySet()) {
-			List<String> values = new ArrayList<>();
-
-			Object object = configurationJSONObject.get(key);
-
-			if (object instanceof JSONArray) {
-				JSONArray jsonArray = configurationJSONObject.getJSONArray(key);
-
-				for (int i = 0; i < jsonArray.length(); i++) {
-					values.add(jsonArray.getString(i));
-				}
-			}
-			else {
-				values.add(String.valueOf(object));
-			}
-
-			configuration.put(key, values.toArray(new String[0]));
-		}
-
-		return configuration;
-	}
-
-	private DefaultLayoutListRetrieverContext
-		_getDefaultLayoutListRetrieverContext(JSONObject collectionJSONObject) {
-
-		DefaultLayoutListRetrieverContext defaultLayoutListRetrieverContext =
-			new DefaultLayoutListRetrieverContext();
-
-		defaultLayoutListRetrieverContext.setConfiguration(
-			_getConfiguration(collectionJSONObject));
-		defaultLayoutListRetrieverContext.setContextObject(
-			Optional.ofNullable(
-				_httpServletRequest.getAttribute(
-					InfoDisplayWebKeys.INFO_LIST_DISPLAY_OBJECT)
-			).orElse(
-				_httpServletRequest.getAttribute(InfoDisplayWebKeys.INFO_ITEM)
-			));
-		defaultLayoutListRetrieverContext.setHttpServletRequest(
-			_httpServletRequest);
-		defaultLayoutListRetrieverContext.setSegmentsEntryIds(
-			_getSegmentsEntryIds());
-
-		return defaultLayoutListRetrieverContext;
-	}
-
 	private long _getFileEntryId(String fieldId, Locale locale)
 		throws Exception {
 
@@ -1219,65 +977,6 @@ public class RenderLayoutStructureDisplayContext {
 		return _frontendTokensJSONObject;
 	}
 
-	private LayoutListRetriever<?, ListObjectReference> _getLayoutListRetriever(
-		JSONObject collectionJSONObject) {
-
-		if ((collectionJSONObject == null) ||
-			(collectionJSONObject.length() <= 0)) {
-
-			return null;
-		}
-
-		LayoutListRetrieverTracker layoutListRetrieverTracker =
-			ServletContextUtil.getLayoutListRetrieverTracker();
-
-		LayoutListRetriever<?, ListObjectReference> layoutListRetriever =
-			(LayoutListRetriever<?, ListObjectReference>)
-				layoutListRetrieverTracker.getLayoutListRetriever(
-					collectionJSONObject.getString("type"));
-
-		if (layoutListRetriever == null) {
-			return null;
-		}
-
-		return layoutListRetriever;
-	}
-
-	private ListObjectReference _getListObjectReference(
-		JSONObject collectionJSONObject) {
-
-		if ((collectionJSONObject == null) ||
-			(collectionJSONObject.length() <= 0)) {
-
-			return null;
-		}
-
-		LayoutListRetrieverTracker layoutListRetrieverTracker =
-			ServletContextUtil.getLayoutListRetrieverTracker();
-
-		String type = collectionJSONObject.getString("type");
-
-		LayoutListRetriever<?, ?> layoutListRetriever =
-			layoutListRetrieverTracker.getLayoutListRetriever(type);
-
-		if (layoutListRetriever == null) {
-			return null;
-		}
-
-		ListObjectReferenceFactoryTracker listObjectReferenceFactoryTracker =
-			ServletContextUtil.getListObjectReferenceFactoryTracker();
-
-		ListObjectReferenceFactory<?> listObjectReferenceFactory =
-			listObjectReferenceFactoryTracker.getListObjectReference(type);
-
-		if (listObjectReferenceFactory == null) {
-			return null;
-		}
-
-		return listObjectReferenceFactory.getListObjectReference(
-			collectionJSONObject);
-	}
-
 	private String _getMainItemId() {
 		if (Validator.isNotNull(_mainItemId)) {
 			return _mainItemId;
@@ -1415,24 +1114,6 @@ public class RenderLayoutStructureDisplayContext {
 		return _previewVersion;
 	}
 
-	private long[] _getSegmentsEntryIds() {
-		if (_segmentsEntryIds != null) {
-			return _segmentsEntryIds;
-		}
-
-		SegmentsEntryRetriever segmentsEntryRetriever =
-			ServletContextUtil.getSegmentsEntryRetriever();
-
-		RequestContextMapper requestContextMapper =
-			ServletContextUtil.getRequestContextMapper();
-
-		_segmentsEntryIds = segmentsEntryRetriever.getSegmentsEntryIds(
-			_themeDisplay.getScopeGroupId(), _themeDisplay.getUserId(),
-			requestContextMapper.map(_httpServletRequest));
-
-		return _segmentsEntryIds;
-	}
-
 	private long[] _getSegmentsExperienceIds() {
 		if (_segmentsExperienceIds != null) {
 			return _segmentsExperienceIds;
@@ -1452,7 +1133,6 @@ public class RenderLayoutStructureDisplayContext {
 	private final Map<String, Object> _fieldValues;
 	private JSONObject _frontendTokensJSONObject;
 	private final HttpServletRequest _httpServletRequest;
-	private final HttpServletResponse _httpServletResponse;
 	private final LayoutStructure _layoutStructure;
 	private final String _mainItemId;
 	private final String _mode;
@@ -1460,7 +1140,6 @@ public class RenderLayoutStructureDisplayContext {
 	private Long _previewClassPK;
 	private Integer _previewType;
 	private String _previewVersion;
-	private long[] _segmentsEntryIds;
 	private long[] _segmentsExperienceIds;
 	private final boolean _showPreview;
 	private final ThemeDisplay _themeDisplay;
